DFL64: small fixes for size_t conversion
This does NOT make any change on Win32 programs.

+++ win32/dfl/control.d
@@ -3647,7 +3647,7 @@ class Control: DObject, IWindow // docmain
 	///
 	static bool isMnemonic(dchar charCode, Dstring text)
 	{
-		uint ui;
+		size_t ui;
 		for(ui = 0; ui != text.length; ui++)
 		{
 			if('&' == text[ui])
